refactor: unify generic expr rewrite functions into the `datafusion_expr::expr_rewriter`
